Convert Client to use `.send` instead of `<<`
As one of the requirements of the Statsd Client, the emitter must be
able to continue even if the server does not exist.

Previously, in order to prevent a crash, we had to create a new socket,
connect, push the message across and close.
When attempting to leave the socket open and push more messages across,
an error was raised regarding no connection.

This implementation allows for a "broadcast-style" emitter, where the
message will be sent to the host,port destination, and if there's a
server listening, great, but otherwise, degrade silently.